# apps/api/src/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Getswyft API
  version: 1.1.0
  description: |
    Express + Socket.IO API for Getswyft.

    Authentication supports:
    - Bearer JWT access tokens for authenticated users
    - `x-api-key` for tenant-scoped external integrations

    Widget visitor routes use their own visitor-scoped bearer token after session creation.
servers:
  - url: https://{host}
    variables:
      host:
        default: api.getswyft.local
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
  /v1/auth/invitations/{token}:
    get:
      tags: [Auth]
      summary: Inspect an invitation token before sign-in or acceptance
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation token status and tenant metadata
        "404":
          description: Invitation not found
...
  /v1/auth/invitations/accept:
    post:
      tags: [Auth]
      summary: Accept an invitation token as the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "201":
          description: Invitation accepted and tenant membership granted
        "403":
          description: Invitation email does not match authenticated user
        "409":
          description: Invitation was revoked or already accepted
        "410":
          description: Invitation expired
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key